Vary colDen by varying nH
Now the depth of the cloud is kept fixed when calculating a CoG,
colDen variations are achieved by alowing nH to vary.

This fits in with the general model I'm employing - the depth of
cloud is fixed within an iteration, but the density within a
cloud is assumed to vary, giving rise to variable column densities.