Add bodge for very thick clouds
With very thick clouds the small wavenumbers are undersampled and
so the integral wrt k_z cannot use the same formulation as when
not undersampled.

Fortunately we know that the fourier transform of the density function
can be approxiamated by a dirac delta function of known integral
in this case and so this approximation is used for very thick clouds.